Separate BLE Related Code into a BLE.cpp
参考
https://github.com/sqfmi/Watchy/blob/342eb48a497eb3dcb1ec03c5c156937fdec07ba9/src/BLE.h

GitHub Code Searchで適当に探したもの。
シンプルだけどわかりやすくていいかなと思った
