+++ src/main.cpp
@@ -2,79 +2,11 @@
 #include "Adafruit_VL53L0X.h"
 #include <Avatar.h>
 #include <unordered_map>
-#include "BLEDevice.h"
-#include "BLEServer.h"
-#include <BLE2902.h>
-
-#define SERVICE_UUID "0B21C05A-44C2-47CC-BFEF-4F7165C33908"
-#define EXPRESSION_CHARACTERISTIC_UUID "B3C450C9-5FC5-48F6-9EFD-D588E494F462"
-#define DISTANCE_CHARACTERISTIC_UUID "29C2D1B2-944A-4FBA-AFCD-133E09532556"
+#include "BLE.h"
 
 using namespace m5avatar;
 
-bool connected = false;
-
-class MyServerCallbacks : public BLEServerCallbacks
-{
-    void onConnect(BLEServer *pServer)
-    {
-        connected = true;
-    };
-
-    void onDisconnect(BLEServer *pServer)
-    {
-        connected = false;
-    }
-};
-
-int expression = static_cast<int>(Expression::Neutral);
-bool isExpressionChanged = false;
-
-class ExpressionCharacteristicCallbacks : public BLECharacteristicCallbacks
-{
-    void onWrite(BLECharacteristic *pCharacteristic)
-    {
-        std::string value = pCharacteristic->getValue();
-        if (value.length() > 0)
-        {
-            if (value[0] == expression)
-            {
-                return;
-            }
-            expression = value[0];
-            isExpressionChanged = true;
-        }
-    }
-};
-
-BLECharacteristic *pDistanceCharacteristic;
-
-void setupServer()
-{
-    BLEDevice::init("M5AtomS3");
-    BLEServer *pServer = BLEDevice::createServer();
-    pServer->setCallbacks(new MyServerCallbacks());
-
-    BLEService *pService = pServer->createService(SERVICE_UUID);
-    BLECharacteristic *pExpressionCharacteristic = pService->createCharacteristic(
-        EXPRESSION_CHARACTERISTIC_UUID,
-        BLECharacteristic::PROPERTY_READ |
-            BLECharacteristic::PROPERTY_WRITE);
-    pExpressionCharacteristic->addDescriptor(new BLE2902());
-    pExpressionCharacteristic->setCallbacks(new ExpressionCharacteristicCallbacks());
-
-    pDistanceCharacteristic = pService->createCharacteristic(
-        DISTANCE_CHARACTERISTIC_UUID,
-        BLECharacteristic::PROPERTY_READ |
-            BLECharacteristic::PROPERTY_NOTIFY);
-    pDistanceCharacteristic->addDescriptor(new BLE2902());
-
-    pService->start();
-
-    BLEAdvertising *pAdvertising = BLEDevice::getAdvertising();
-    pAdvertising->addServiceUUID(SERVICE_UUID);
-    pAdvertising->setScanResponse(true);
-}
+BLE ble = BLE();
 
 Adafruit_VL53L0X lox = Adafruit_VL53L0X();
 
@@ -119,10 +51,17 @@ public:
     virtual void onExit() {}
 };
 
-State currentState;
-State nextState;
+State currentState = State::Start;
+State nextState = State::Start;
 BaseStateBehavior *pStateBehavior;
 std::unordered_map<State, BaseStateBehavior *> stateBehaviors;
+
+void registerBehaviors()
+{
+    stateBehaviors[State::Start] = new StartStateBehavior();
+    stateBehaviors[State::Ready] = new ReadyStateBehavior();
+    stateBehaviors[State::Idle] = new IdleStateBehavior();
+}
 
 void setNextState(State state)
 {
@@ -161,7 +100,7 @@ public:
     void onExit() override
     {
         setupAvatar();
-        setupServer();
+        ble.start();
         setupLox();
     }
 };
@@ -176,7 +115,7 @@ public:
 
     void onUpdate() override
     {
-        if (connected)
+        if (ble.isConnected())
         {
             setNextState(State::Idle);
         }
@@ -194,21 +133,6 @@ private:
     unsigned long latestReadRangeTime = 0;
     const unsigned long readRangeInterval = 100;
 
-    std::string distanceNotifyData(uint16_t value)
-    {
-        std::string data;
-        data += static_cast<char>(value & 0xff);
-        data += static_cast<char>((value >> 8) & 0xff);
-        return data;
-    }
-
-    void readDistance()
-    {
-        uint16_t distance = lox.readRange();
-        pDistanceCharacteristic->setValue(distanceNotifyData(distance));
-        pDistanceCharacteristic->notify();
-    }
-
 public:
     void onEnter() override
     {
@@ -217,22 +141,22 @@ public:
 
     void onUpdate() override
     {
-        if (!connected)
+        if (!ble.isConnected())
         {
             setNextState(State::Ready);
             return;
         }
 
-        if (isExpressionChanged)
+        if (ble.isExpressionChanged())
         {
-            avatar.setExpression(static_cast<Expression>(expression));
-            isExpressionChanged = false;
+            avatar.setExpression(static_cast<Expression>(ble.getExpression()));
         }
 
         unsigned long now = millis();
         if (now - latestReadRangeTime >= readRangeInterval && lox.isRangeComplete())
         {
-            readDistance();
+            uint16_t distance = lox.readRange();
+            ble.sendDistance(distance);
             latestReadRangeTime = now;
         }
     }
@@ -248,13 +172,8 @@ void setup()
     auto cfg = M5.config();
     M5.begin(cfg);
     USBSerial.begin(115200);
-    currentState = State::Start;
-    nextState = State::Start;
-    StartStateBehavior *pStartStateBehavior = new StartStateBehavior();
-    pStateBehavior = pStartStateBehavior;
-    stateBehaviors[State::Start] = pStartStateBehavior;
-    stateBehaviors[State::Ready] = new ReadyStateBehavior();
-    stateBehaviors[State::Idle] = new IdleStateBehavior();
+    registerBehaviors();
+    pStateBehavior = stateBehaviors[State::Start];
     pStateBehavior->onEnter();
 }
